pyproject.toml
[tool.poetry]
name = "mugbot"
version = "0.1.0"
description = ""
authors = ["Skynet0 <ethan.ordentlich2017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.5.1"}
"discord.py-stubs" = "^1.5.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
